+++ pkg/api/handlers/compat/unsupported.go
@@ -4,9 +4,8 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/containers/podman/v2/pkg/domain/entities"
-
 	"github.com/containers/podman/v2/pkg/api/handlers/utils"
+	"github.com/containers/podman/v2/pkg/errorhandling"
 	log "github.com/sirupsen/logrus"
 )
 
@@ -14,5 +13,5 @@ func UnsupportedHandler(w http.ResponseWriter, r *http.Request) {
 	msg := fmt.Sprintf("Path %s is not supported", r.URL.Path)
 	log.Infof("Request Failed: %s", msg)
 
-	utils.WriteJSON(w, http.StatusNotFound, entities.ErrorModel{Message: msg})
+	utils.WriteJSON(w, http.StatusNotFound, errorhandling.ErrorModel{Message: msg})
 }
